@@ -8,7 +8,6 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Yarp.ReverseProxy.Configuration;
-using Yarp.ReverseProxy.Forwarder;
 using Yarp.ReverseProxy.Model;
 
 namespace Yarp.Sample
@@ -27,12 +26,6 @@ namespace Yarp.Sample
         /// </summary>
         public void ConfigureServices(IServiceCollection services)
         {
-#if !NET6_0_OR_GREATER
-            // Workaround the lack of distributed tracing support in SocketsHttpHandler before .NET 6.0
-            services.AddSingleton<IForwarderHttpClientFactory, DiagnosticsHandlerFactory>();
-#endif
-
-            // Specify a custom proxy config provider, in this case defined in InMemoryConfigProvider.cs
             // Programatically creating route and cluster configs. This allows loading the data from an arbitrary source.
             services.AddReverseProxy()
                 .LoadFromMemory(GetRoutes(), GetClusters());
@@ -46,6 +39,11 @@ namespace Yarp.Sample
             app.UseRouting();
             app.UseEndpoints(endpoints =>
             {
+                endpoints.Map("/update", context =>
+                {
+                    context.RequestServices.GetRequiredService<InMemoryConfigProvider>().Update(GetRoutes(), GetClusters());
+                    return Task.CompletedTask;
+                });
                 // We can customize the proxy pipeline and add/remove/replace steps
                 endpoints.MapReverseProxy(proxyPipeline =>
                 {
@@ -64,7 +62,7 @@ namespace Yarp.Sample
             {
                 new RouteConfig()
                 {
-                    RouteId = "route1",
+                    RouteId = "route" + Random.Shared.Next(), // Forces a new route id each time GetRoutes is called.
                     ClusterId = "cluster1",
                     Match = new RouteMatch
                     {
