@@ -3,7 +3,6 @@
 
 using k8s;
 using k8s.Models;
-using Microsoft.Kubernetes;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using Yarp.Kubernetes.Controller.Services;
@@ -20,6 +19,7 @@ public interface ICache
     bool Update(WatchEventType eventType, V1Ingress ingress);
     ImmutableList<string> Update(WatchEventType eventType, V1Service service);
     ImmutableList<string> Update(WatchEventType eventType, V1Endpoints endpoints);
+    void Update(WatchEventType eventType, V1Secret secret);
     bool TryGetReconcileData(NamespacedName key, out ReconcileData data);
     void GetKeys(List<NamespacedName> keys);
     IEnumerable<IngressData> GetIngresses();
