@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Yarp.ReverseProxy.Configuration;
 
-namespace Yarp.Kubernetes.Controller.Services;
+namespace Yarp.Kubernetes.Controller.Converters;
 
 internal class YarpConfigContext
 {
