@@ -7,7 +7,6 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Microsoft.Kubernetes;
 using Yarp.Kubernetes.Controller.Caching;
 using Yarp.Kubernetes.Controller.Configuration;
 using Yarp.Kubernetes.Controller.Converters;
@@ -41,10 +40,17 @@ public partial class Reconciler : IReconciler
 
             foreach (var ingress in ingresses)
             {
-                if (_cache.TryGetReconcileData(new NamespacedName(ingress.Metadata.NamespaceProperty, ingress.Metadata.Name), out var data))
+                try
                 {
-                    var ingressContext = new YarpIngressContext(ingress, data.ServiceList, data.EndpointsList);
-                    YarpParser.ConvertFromKubernetesIngress(ingressContext, configContext);
+                    if (_cache.TryGetReconcileData(new NamespacedName(ingress.Metadata.NamespaceProperty, ingress.Metadata.Name), out var data))
+                    {
+                        var ingressContext = new YarpIngressContext(ingress, data.ServiceList, data.EndpointsList);
+                        YarpParser.ConvertFromKubernetesIngress(ingressContext, configContext);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    _logger.LogWarning(ex, "Uncaught exception occured while reconciling ingress {IngressNamespace}/{IngressName}", ingress.Metadata.NamespaceProperty, ingress.Metadata.Name);
                 }
             }
 
@@ -57,7 +63,7 @@ public partial class Reconciler : IReconciler
         }
         catch (Exception ex)
         {
-            _logger.LogWarning(ex.Message);
+            _logger.LogWarning(ex, "Uncaught exception occured while reconciling");
             throw;
         }
     }
