@@ -14,7 +14,6 @@ using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Features;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Primitives;
 using Microsoft.Net.Http.Headers;
 using Yarp.ReverseProxy.Utilities;
 
@@ -27,9 +26,7 @@ internal sealed class HttpForwarder : IHttpForwarder
 {
     private static readonly TimeSpan DefaultTimeout = TimeSpan.FromSeconds(100);
     private static readonly Version DefaultVersion = HttpVersion.Version20;
-#if NET
     private static readonly HttpVersionPolicy DefaultVersionPolicy = HttpVersionPolicy.RequestVersionOrLower;
-#endif
     private readonly ILogger _logger;
     private readonly IClock _clock;
 
@@ -271,13 +268,11 @@ internal sealed class HttpForwarder : IHttpForwarder
 
         // Default to HTTP/1.1 for proxying upgradeable requests. This is already the default as of .NET Core 3.1
         // Otherwise request what's set in proxyOptions (e.g. default HTTP/2) and let HttpClient negotiate the protocol
-        // based on VersionPolicy (for .NET 5 and higher). For example, downgrading to HTTP/1.1 if it cannot establish HTTP/2 with the target.
+        // based on VersionPolicy. For example, downgrading to HTTP/1.1 if it cannot establish HTTP/2 with the target.
         // This is done without extra round-trips thanks to ALPN. We can detect a downgrade after calling HttpClient.SendAsync
         // (see Step 3 below). TBD how this will change when HTTP/3 is supported.
         destinationRequest.Version = isUpgradeRequest ? ProtocolHelper.Http11Version : (requestConfig?.Version ?? DefaultVersion);
-#if NET
         destinationRequest.VersionPolicy = isUpgradeRequest ? HttpVersionPolicy.RequestVersionOrLower : (requestConfig?.VersionPolicy ?? DefaultVersionPolicy);
-#endif
 
         // :: Step 2: Setup copy of request body (background) Client --► Proxy --► Destination
         // Note that we must do this before step (3) because step (3) may also add headers to the HttpContent that we set up here.
@@ -337,15 +332,12 @@ internal sealed class HttpForwarder : IHttpForwarder
         var contentLength = request.Headers.ContentLength;
         var method = request.Method;
 
-#if NET
         var canHaveBodyFeature = request.HttpContext.Features.Get<IHttpRequestBodyDetectionFeature>();
         if (canHaveBodyFeature is not null)
         {
             // 5.0 servers provide a definitive answer for us.
             hasBody = canHaveBodyFeature.CanHaveBody;
         }
-        else
-#endif
         // https://tools.ietf.org/html/rfc7230#section-3.3.3
         // All HTTP/1.1 requests should have Transfer-Encoding or Content-Length.
         // Http.Sys/IIS will even add a Transfer-Encoding header to HTTP/2 requests with bodies for back-compat.
@@ -353,7 +345,7 @@ internal sealed class HttpForwarder : IHttpForwarder
         // https://tools.ietf.org/html/rfc1945#section-7.2.2
         //
         // Transfer-Encoding overrides Content-Length per spec
-        if (request.Headers.TryGetValue(HeaderNames.TransferEncoding, out var transferEncoding)
+        else if (request.Headers.TryGetValue(HeaderNames.TransferEncoding, out var transferEncoding)
             && transferEncoding.Count == 1
             && string.Equals("chunked", transferEncoding.ToString(), StringComparison.OrdinalIgnoreCase))
         {
@@ -462,11 +454,7 @@ internal sealed class HttpForwarder : IHttpForwarder
             }
             else
             {
-#if NET6_0_OR_GREATER
                 Debug.Assert(requestCancellationSource.IsCancellationRequested || requestException.ToString().Contains("ConnectTimeout"), requestException.ToString());
-#else
-                Debug.Assert(requestCancellationSource.IsCancellationRequested || requestException.ToString().Contains("ConnectHelper"), requestException.ToString());
-#endif
                 return await ReportErrorAsync(ForwarderError.RequestTimedOut, StatusCodes.Status504GatewayTimeout);
             }
         }
@@ -522,18 +510,17 @@ internal sealed class HttpForwarder : IHttpForwarder
 
     private static void RestoreUpgradeHeaders(HttpContext context, HttpResponseMessage response)
     {
-        // We don't use RequestUtilities.TryGetValues for the Connection as we do want value validation.
+        // We don't use NonValidated for the Connection header as we do want value validation.
         // HttpHeaders.TryGetValues will handle the parsing and split the values for us.
         if (RequestUtilities.TryGetValues(response.Headers, HeaderNames.Upgrade, out var upgradeValues)
             && response.Headers.TryGetValues(HeaderNames.Connection, out var connectionValues))
         {
-            context.Response.Headers.TryAdd(HeaderNames.Upgrade, upgradeValues);
-
             foreach (var value in connectionValues)
             {
                 if (value.Equals("upgrade", StringComparison.OrdinalIgnoreCase))
                 {
                     context.Response.Headers.TryAdd(HeaderNames.Connection, value);
+                    context.Response.Headers.TryAdd(HeaderNames.Upgrade, upgradeValues);
                     break;
                 }
             }
@@ -580,7 +567,7 @@ internal sealed class HttpForwarder : IHttpForwarder
         using var clientStream = upgradeResult;
 
         // :: Step 7-A-2: Copy duplex streams
-        using var destinationStream = await destinationResponse.Content.ReadAsStreamAsync();
+        using var destinationStream = await destinationResponse.Content.ReadAsStreamAsync(activityCancellationSource.Token);
 
         var requestTask = StreamCopier.CopyAsync(isRequest: true, clientStream, destinationStream, StreamCopier.UnknownLength, _clock, activityCancellationSource, activityCancellationSource.Token).AsTask();
         var responseTask = StreamCopier.CopyAsync(isRequest: false, destinationStream, clientStream, StreamCopier.UnknownLength, _clock, activityCancellationSource, activityCancellationSource.Token).AsTask();
@@ -639,7 +626,7 @@ internal sealed class HttpForwarder : IHttpForwarder
         // https://github.com/dotnet/runtime/blame/8fc68f626a11d646109a758cb0fc70a0aa7826f1/src/libraries/System.Net.Http/src/System/Net/Http/HttpResponseMessage.cs#L46
         if (destinationResponseContent is not null)
         {
-            using var destinationResponseStream = await destinationResponseContent.ReadAsStreamAsync();
+            using var destinationResponseStream = await destinationResponseContent.ReadAsStreamAsync(activityCancellationSource.Token);
             // The response content-length is enforced by the server.
             return await StreamCopier.CopyAsync(isRequest: false, destinationResponseStream, clientResponseStream, StreamCopier.UnknownLength, _clock, activityCancellationSource, activityCancellationSource.Token);
         }
@@ -782,12 +769,8 @@ internal sealed class HttpForwarder : IHttpForwarder
             if (logger.IsEnabled(LogLevel.Information))
             {
                 var streaming = isStreamingRequest ? "streaming" : "no-streaming";
-                var version = ProtocolHelper.GetHttpProtocol(msg.Version);
-#if NET
+                var version = HttpProtocol.GetHttpProtocol(msg.Version);
                 var versionPolicy = ProtocolHelper.GetVersionPolicy(msg.VersionPolicy);
-#else
-                var versionPolicy = "RequestVersionOrLower";
-#endif
                 _proxying(logger, msg.RequestUri!.AbsoluteUri, version, versionPolicy, streaming, null);
             }
         }
