@@ -3,13 +3,13 @@
 
 using System;
 using System.Diagnostics;
-using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
+using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
-using Microsoft.Extensions.Primitives;
 using Microsoft.Net.Http.Headers;
 using Yarp.ReverseProxy.Transforms.Builder;
 
@@ -33,6 +33,25 @@ public class HttpTransformer
     /// Used to create derived instances.
     /// </summary>
     protected HttpTransformer() { }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private static bool IsBodylessStatusCode(HttpStatusCode statusCode) =>
+        statusCode switch
+        {
+            // A 1xx response is terminated by the end of the header section; it cannot contain content
+            // or trailers.
+            // See https://www.rfc-editor.org/rfc/rfc9110.html#section-15.2-2
+            >= HttpStatusCode.Continue and < HttpStatusCode.OK => true,
+            // A 204 response is terminated by the end of the header section; it cannot contain content
+            // or trailers.
+            // See https://www.rfc-editor.org/rfc/rfc9110.html#section-15.3.5-5
+            HttpStatusCode.NoContent => true,
+            // Since the 205 status code implies that no additional content will be provided, a server
+            // MUST NOT generate content in a 205 response.
+            // See https://www.rfc-editor.org/rfc/rfc9110.html#section-15.3.6-3
+            HttpStatusCode.ResetContent => true,
+            _ => false
+        };
 
     /// <summary>
     /// A callback that is invoked prior to sending the proxied request. All HttpRequestMessage fields are
@@ -130,8 +149,18 @@ public class HttpTransformer
         // remove the received Content-Length field prior to forwarding such
         // a message downstream.
         if (proxyResponse.Content is not null
-            && RequestUtilities.ContainsHeader(proxyResponse.Headers, HeaderNames.TransferEncoding)
-            && RequestUtilities.ContainsHeader(proxyResponse.Content.Headers, HeaderNames.ContentLength))
+            && proxyResponse.Headers.NonValidated.Contains(HeaderNames.TransferEncoding)
+            && proxyResponse.Content.Headers.NonValidated.Contains(HeaderNames.ContentLength))
+        {
+            httpContext.Response.Headers.Remove(HeaderNames.ContentLength);
+        }
+
+        // For responses with status codes that shouldn't include a body,
+        // we remove the 'Content-Length: 0' header if one is present.
+        if (proxyResponse.Content is not null
+            && IsBodylessStatusCode(proxyResponse.StatusCode)
+            && proxyResponse.Content.Headers.NonValidated.TryGetValues(HeaderNames.ContentLength, out var contentLengthValue)
+            && contentLengthValue.ToString() == "0")
         {
             httpContext.Response.Headers.Remove(HeaderNames.ContentLength);
         }
@@ -166,7 +195,6 @@ public class HttpTransformer
     {
         // We want to append to any prior values, if any.
         // Not using Append here because it skips empty headers.
-#if NET6_0_OR_GREATER
         foreach (var header in source.NonValidated)
         {
             var headerName = header.Key;
@@ -177,19 +205,5 @@ public class HttpTransformer
 
             destination[headerName] = RequestUtilities.Concat(destination[headerName], header.Value);
         }
-#else
-        foreach (var header in source)
-        {
-            var headerName = header.Key;
-            if (RequestUtilities.ShouldSkipResponseHeader(headerName))
-            {
-                continue;
-            }
-
-            Debug.Assert(header.Value is string[]);
-            var values = header.Value as string[] ?? header.Value.ToArray();
-            destination[headerName] = StringValues.Concat(destination[headerName], values);
-        }
-#endif
     }
 }
