@@ -14,7 +14,6 @@ using k8s;
 using k8s.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Microsoft.Kubernetes;
 using Moq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
@@ -22,15 +21,15 @@ using Newtonsoft.Json.Linq;
 using Xunit;
 using Yarp.Kubernetes.Controller;
 using Yarp.Kubernetes.Controller.Caching;
+using Yarp.Kubernetes.Controller.Certificates;
 using Yarp.Kubernetes.Controller.Converters;
-using Yarp.Kubernetes.Controller.Services;
 using JsonSerializer = System.Text.Json.JsonSerializer;
 
 namespace Yarp.Kubernetes.Tests;
 
-public class IngressControllerTests
+public class IngressConversionTests
 {
-    public IngressControllerTests()
+    public IngressConversionTests()
     {
         JsonConvert.DefaultSettings = () => new JsonSerializerSettings() {
             NullValueHandling = NullValueHandling.Ignore,
@@ -45,11 +44,13 @@ public class IngressControllerTests
     [InlineData("exact-match")]
     [InlineData("annotations")]
     [InlineData("mapped-port")]
+    [InlineData("port-mismatch")]
     [InlineData("hostname-routing")]
     [InlineData("multiple-ingresses")]
     [InlineData("multiple-ingresses-one-svc")]
     [InlineData("multiple-namespaces")]
     [InlineData("route-metadata")]
+    [InlineData("missing-svc")]
     public async Task ParsingTests(string name)
     {
         var ingressClass = KubeResourceGenerator.CreateIngressClass("yarp", "microsoft.com/ingress-yarp", true);
@@ -80,30 +81,31 @@ public class IngressControllerTests
         VerifyJson(routesJson, name, "routes.json");
     }
 
-        private static string StripNullProperties(string json)
+    private static string StripNullProperties(string json)
+    {
+        using var reader = new JsonTextReader(new StringReader(json));
+        var sb = new StringBuilder();
+        using var sw = new StringWriter(sb);
+        using var writer = new JsonTextWriter(sw);
+        while (reader.Read())
         {
-            using var reader = new JsonTextReader(new StringReader(json));
-            var sb = new StringBuilder();
-            using var sw = new StringWriter(sb);
-            using var writer = new JsonTextWriter(sw);
-            while (reader.Read())
+            var token = reader.TokenType;
+            var value = reader.Value;
+            if(reader.TokenType == JsonToken.PropertyName)
             {
-                var token = reader.TokenType;
-                var value = reader.Value;
-                if(reader.TokenType == JsonToken.PropertyName)
+                reader.Read();
+                if (reader.TokenType == JsonToken.Null)
                 {
-                    reader.Read();
-                    if (reader.TokenType == JsonToken.Null)
-                    {
-                        continue;
-                    }
-                    writer.WriteToken(token, value);
+                    continue;
                 }
-                writer.WriteToken(reader.TokenType, reader.Value);
+                writer.WriteToken(token, value);
             }
-
-            return sb.ToString();
+            writer.WriteToken(reader.TokenType, reader.Value);
         }
+
+        return sb.ToString();
+    }
+
     private static void VerifyJson(string json, string name, string fileName)
     {
         var other = File.ReadAllText(Path.Combine("testassets", name, fileName));
@@ -119,10 +121,13 @@ public class IngressControllerTests
     {
         var mockLogger = new Mock<ILogger<IngressCache>>();
         var mockOptions = new Mock<IOptions<YarpOptions>>();
+        var certificateSelector = new Mock<IServerCertificateSelector>();
+        var loggerHelper = new Mock<ILogger<CertificateHelper>>();
+        var certificateHelper = new CertificateHelper(loggerHelper.Object);
 
         mockOptions.SetupGet(o => o.Value).Returns(new YarpOptions { ControllerClass = "microsoft.com/ingress-yarp" });
 
-        var cache = new IngressCache(mockOptions.Object, mockLogger.Object);
+        var cache = new IngressCache(mockOptions.Object, certificateSelector.Object, certificateHelper, mockLogger.Object);
 
         var typeMap = new Dictionary<string, Type>();
         typeMap.Add("networking.k8s.io/v1/Ingress", typeof(V1Ingress));
