@@ -1,12 +1,12 @@
-﻿// Copyright (c) Microsoft Corporation.
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
 using k8s.Models;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Kubernetes.Testing.Models;
 using System.Threading.Tasks;
+using Yarp.Kubernetes.Tests.TestCluster.Models;
 
-namespace Microsoft.Kubernetes.Testing;
+namespace Yarp.Kubernetes.Tests.TestCluster;
 
 [Route("apis/{group}/{version}/{plural}")]
 public class ResourceApiGroupController : ControllerBase
