@@ -48,22 +48,18 @@ public class TelemetryConsumptionTests
             services.AddSingleton<TelemetryConsumer>();
             services.AddSingleton(services => (IForwarderTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
             services.AddSingleton(services => (IKestrelTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
-#if NET
             services.AddSingleton(services => (IHttpTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
             services.AddSingleton(services => (INameResolutionTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
             services.AddSingleton(services => (INetSecurityTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
             services.AddSingleton(services => (ISocketsTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
-#endif
 
             services.AddSingleton<SecondTelemetryConsumer>();
             services.AddSingleton(services => (IForwarderTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
             services.AddSingleton(services => (IKestrelTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
-#if NET
             services.AddSingleton(services => (IHttpTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
             services.AddSingleton(services => (INameResolutionTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
             services.AddSingleton(services => (INetSecurityTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
             services.AddSingleton(services => (ISocketsTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
-#endif
 
             services.AddTelemetryListeners();
         }
@@ -105,22 +101,18 @@ public class TelemetryConsumptionTests
             "OnForwarderInvoke",
             "OnForwarderStart",
             "OnForwarderStage-SendAsyncStart",
-#if NET
             "OnRequestStart",
             "OnConnectStart",
             "OnConnectStop",
             "OnHandshakeStart",
             "OnHandshakeStop",
             "OnConnectionEstablished",
-#if NET6_0_OR_GREATER
             "OnRequestLeftQueue",
-#endif
             "OnRequestHeadersStart",
             "OnRequestHeadersStop",
             "OnResponseHeadersStart",
             "OnResponseHeadersStop",
             "OnRequestStop",
-#endif
             "OnForwarderStage-SendAsyncStop",
             "OnForwarderStage-ResponseContentTransferStart",
             "OnContentTransferred",
@@ -135,7 +127,6 @@ public class TelemetryConsumptionTests
         }
     }
 
-#if NET
     [Theory]
     [InlineData(RegistrationApproach.WithInstanceHelper)]
     [InlineData(RegistrationApproach.WithGenericHelper)]
@@ -162,9 +153,7 @@ public class TelemetryConsumptionTests
             "OnConnectStart",
             "OnConnectStop",
             "OnConnectionEstablished",
-#if NET6_0_OR_GREATER
             "OnRequestLeftQueue",
-#endif
             "OnRequestHeadersStart",
             "OnRequestHeadersStop",
             "OnResponseHeadersStart",
@@ -178,20 +167,16 @@ public class TelemetryConsumptionTests
             VerifyStages(expected, stages);
         }
     }
-#endif
 
     private class SecondTelemetryConsumer : TelemetryConsumer { }
 
     private class TelemetryConsumer :
         IForwarderTelemetryConsumer,
-        IKestrelTelemetryConsumer
-#if NET
-        ,
+        IKestrelTelemetryConsumer,
         IHttpTelemetryConsumer,
         INameResolutionTelemetryConsumer,
         INetSecurityTelemetryConsumer,
         ISocketsTelemetryConsumer
-#endif
     {
         public static readonly ConcurrentDictionary<(string, Type), List<(string Stage, DateTime Timestamp)>> PerClusterTelemetry = new();
         public static readonly ConcurrentDictionary<(string, Type), List<(string Stage, DateTime Timestamp)>> PerPathAndQueryTelemetry = new();
@@ -219,7 +204,6 @@ public class TelemetryConsumptionTests
             AddStage(nameof(OnForwarderInvoke), timestamp);
             PerClusterTelemetry.TryAdd((clusterId, GetType()), _stages.Value);
         }
-#if NET
         public void OnRequestStart(DateTime timestamp, string scheme, string host, int port, string pathAndQuery, int versionMajor, int versionMinor, HttpVersionPolicy versionPolicy)
         {
             AddStage(nameof(OnRequestStart), timestamp);
@@ -246,10 +230,6 @@ public class TelemetryConsumptionTests
         public void OnConnectFailed(DateTime timestamp, SocketError error, string exceptionMessage) => AddStage(nameof(OnConnectFailed), timestamp);
         public void OnRequestStart(DateTime timestamp, string connectionId, string requestId, string httpVersion, string path, string method) => AddStage($"{nameof(OnRequestStart)}-Kestrel", timestamp);
         public void OnRequestStop(DateTime timestamp, string connectionId, string requestId, string httpVersion, string path, string method) => AddStage($"{nameof(OnRequestStop)}-Kestrel", timestamp);
-#else
-        public void OnRequestStart(DateTime timestamp, string connectionId, string requestId) => AddStage($"{nameof(OnRequestStart)}-Kestrel", timestamp);
-        public void OnRequestStop(DateTime timestamp, string connectionId, string requestId) => AddStage($"{nameof(OnRequestStop)}-Kestrel", timestamp);
-#endif
     }
 
     [Fact]
@@ -269,13 +249,11 @@ public class TelemetryConsumptionTests
                 var services = proxyBuilder.Services;
 
                 services.AddSingleton<IMetricsConsumer<ForwarderMetrics>>(consumer);
-#if NET
                 services.AddSingleton<IMetricsConsumer<KestrelMetrics>>(consumer);
                 services.AddSingleton<IMetricsConsumer<HttpMetrics>>(consumer);
                 services.AddSingleton<IMetricsConsumer<SocketsMetrics>>(consumer);
                 services.AddSingleton<IMetricsConsumer<NetSecurityMetrics>>(consumer);
                 services.AddSingleton<IMetricsConsumer<NameResolutionMetrics>>(consumer);
-#endif
 
                 services.AddTelemetryListeners();
             },
@@ -295,26 +273,20 @@ public class TelemetryConsumptionTests
             catch { }
 
             await Task.WhenAll(
-                WaitAsync(() => consumer.ProxyMetrics.LastOrDefault()?.RequestsStarted > 0, nameof(ForwarderMetrics))
-#if NET
-                ,
+                WaitAsync(() => consumer.ProxyMetrics.LastOrDefault()?.RequestsStarted > 0, nameof(ForwarderMetrics)),
                 WaitAsync(() => consumer.KestrelMetrics.LastOrDefault()?.TotalConnections > 0, nameof(KestrelMetrics)),
                 WaitAsync(() => consumer.HttpMetrics.LastOrDefault()?.RequestsStarted > 0, nameof(HttpMetrics)),
                 WaitAsync(() => consumer.SocketsMetrics.LastOrDefault()?.OutgoingConnectionsEstablished > 0, nameof(SocketsMetrics)),
                 WaitAsync(() => consumer.NetSecurityMetrics.LastOrDefault()?.TotalTlsHandshakes > 0, nameof(NetSecurityMetrics)),
-                WaitAsync(() => consumer.NameResolutionMetrics.LastOrDefault()?.DnsLookupsRequested > 0, nameof(NameResolutionMetrics))
-#endif
-                );
+                WaitAsync(() => consumer.NameResolutionMetrics.LastOrDefault()?.DnsLookupsRequested > 0, nameof(NameResolutionMetrics)));
         });
 
         VerifyTimestamp(consumer.ProxyMetrics.Last().Timestamp);
-#if NET
         VerifyTimestamp(consumer.KestrelMetrics.Last().Timestamp);
         VerifyTimestamp(consumer.HttpMetrics.Last().Timestamp);
         VerifyTimestamp(consumer.SocketsMetrics.Last().Timestamp);
         VerifyTimestamp(consumer.NetSecurityMetrics.Last().Timestamp);
         VerifyTimestamp(consumer.NameResolutionMetrics.Last().Timestamp);
-#endif
 
         static void VerifyTimestamp(DateTime timestamp)
         {
@@ -337,32 +309,25 @@ public class TelemetryConsumptionTests
     }
 
     private sealed class MetricsConsumer :
-        IMetricsConsumer<ForwarderMetrics>
-#if NET
-        ,
+        IMetricsConsumer<ForwarderMetrics>,
         IMetricsConsumer<KestrelMetrics>,
         IMetricsConsumer<HttpMetrics>,
         IMetricsConsumer<NameResolutionMetrics>,
         IMetricsConsumer<NetSecurityMetrics>,
         IMetricsConsumer<SocketsMetrics>
-#endif
     {
         public readonly ConcurrentQueue<ForwarderMetrics> ProxyMetrics = new ConcurrentQueue<ForwarderMetrics>();
-#if NET
         public readonly ConcurrentQueue<KestrelMetrics> KestrelMetrics = new();
         public readonly ConcurrentQueue<HttpMetrics> HttpMetrics = new();
         public readonly ConcurrentQueue<SocketsMetrics> SocketsMetrics = new();
         public readonly ConcurrentQueue<NetSecurityMetrics> NetSecurityMetrics = new();
         public readonly ConcurrentQueue<NameResolutionMetrics> NameResolutionMetrics = new();
-#endif
 
         public void OnMetrics(ForwarderMetrics previous, ForwarderMetrics current) => ProxyMetrics.Enqueue(current);
-#if NET
         public void OnMetrics(KestrelMetrics previous, KestrelMetrics current) => KestrelMetrics.Enqueue(current);
         public void OnMetrics(SocketsMetrics previous, SocketsMetrics current) => SocketsMetrics.Enqueue(current);
         public void OnMetrics(NetSecurityMetrics previous, NetSecurityMetrics current) => NetSecurityMetrics.Enqueue(current);
         public void OnMetrics(NameResolutionMetrics previous, NameResolutionMetrics current) => NameResolutionMetrics.Enqueue(current);
         public void OnMetrics(HttpMetrics previous, HttpMetrics current) => HttpMetrics.Enqueue(current);
-#endif
     }
 }
